Set Deployment Target to iOS 8, Fixed Warnings

+++ NTPKit/NTPServer.m
@@ -29,7 +29,7 @@ static const uint32_t kSecondsFrom1900To1970 = 2208988800UL;
 static ufixed64_t ntp_localtime_get_ufixed64() {
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return ufixed64(tv.tv_sec + kSecondsFrom1900To1970, tv.tv_usec * (pow(2, 32) / USEC_PER_SEC));
+    return ufixed64((uint32_t)tv.tv_sec + kSecondsFrom1900To1970, tv.tv_usec * (pow(2, 32) / USEC_PER_SEC));
 }
 
 + (NTPServer *)defaultServer {
